FLYPIE-176: Add missing field to facets.
This is a better alternative to #118.

It handles all cases for missing field.
And it handles the errors at the data level.

It properly handles adding and removing missing field facet.

We could still look to do the data stuff on funnel_cake_index or the
solr schema.

But this suffices for now.

* wip - missing specs.