FLYPIE-271: Fix fetch record fails.
The RealTime Get feature for the funcake oai collection is broken and
it's not obvious why this is happening.

This is hopefully a temporary work around that uses a `select?fq=id:<id>` to fetch single records.

The `select` search handler uses a less efficient process for getting
single record so it would be good if we can fix RealTime Get on the OAI
collections.